+++ api/src/main/java/org/allaymc/api/world/WorldViewer.java
@@ -56,15 +56,6 @@ public interface WorldViewer extends DebugShapeViewer {
     void viewEntityLocation(Entity entity, Location3dc newLocation, boolean teleporting);
 
     /**
-     * @deprecated Please use {@link #viewEntityLocation(Entity, Location3dc, boolean)} instead.
-     */
-    @Deprecated(since = "0.24.0", forRemoval = true)
-    @ApiStatus.ScheduledForRemoval(inVersion = "0.26.0")
-    default void viewEntityLocation(Entity entity, Location3d locationLastSent, Location3dc newLocation, boolean teleporting) {
-        viewEntityLocation(entity, newLocation, teleporting);
-    }
-
-    /**
      * Views the motion of the entity passed.
      *
      * @param entity the entity to view